The code is cleaner and safer than before and added more tests
execute(constStr input) added which removes the clutter to remember how to
initialize and execute an expression.

Now most of the arguments are of constStr type to make reusing functions
safer. If some are not then either they have to be unsafe or they have been
overlooked.

Println - depends on isQuiet to print or not
println - Will always print with a newline at the end

Removed usage of '\n' which isn't cross platform

Prints more intelligently

More tests need to be added