Cleaned GUI files to contain the necessary source files only
Also fixed a bug by adding a feature to store or not to store the answers.

The bug: Whenever the answer was parallely calculated from the GUI, the
resulting answers were always stored which was not the desired behaviour.
Instead it was better to store answers only when return was pressed.

Now `make all_gui` can be invoked to build the Qt GUI and produce the target
executable. No need to worry about qmake. By default `make` will not compile
the GUI executables.