Fixed a glitch which cleared the input when return was pressed
Expressions and their answers should only be recorded when it yields an
answer. If it results in some arbitrary error or no output then the user has to
fix it else the input will not be recorded. But when the user pressed return in
such a case, it was good that we hadn't recorded the answer but the input field
became empty which was a bad idea as the user might have written some valuable
information in it.

This was caused because the tab order of the clearButton was just after the
lineEditInput. Changing the tab order to be calculateButton after lineEditInput
fixed the problem.